containers/tesseract-ocr/requirements.txt
Flask>=2.3.0
Pillow>=10.0.0
pytesseract>=0.3.0
pdf2image>=1.16.0
